# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Multiplayer Time API
# during dev, should point to your local machine
host: localhost:2000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  jsonWebToken:
    type: apiKey
    in: header
    name: Authorization
  jsonAdminWebToken:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /user:
    x-swagger-router-controller: user_controller
    get:
      operationId: getAllUsers
      tags:
        - user
      description: get User information
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: addUser
      tags: 
        - user
      description: replace user info
      parameters:
        - name: User
          description: User properties
          in: body
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/avatar/:
    x-swagger-router-controller: user_controller
    post:
      operationId: updateAvatar
      tags: 
        - user
      security:
        - jsonWebToken: []
      description: replace user avatar
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: avatar 
          type: file
          description: The file to replace.
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse" 
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/password/:
    x-swagger-router-controller: user_controller
    post:
      operationId: updatePassword 
      tags: 
        - user
      description: replace user avatar
      parameters:
        - name: User
          description: User properties
          in: body
          required: true
          schema:
            type: object
            properties:
              password:
                type: string
              newpassword: 
               type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse" 
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{id}:
    x-swagger-router-controller: user_controller
    get:
      operationId: getProfile
      description: get a User
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserProfile"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/username/{id}:
    x-swagger-router-controller: user_controller
    get:
      operationId: checkUsername
      tags:
        - user
      description: update a User
      parameters:
        - name: id
          description: User id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse" 
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/email/{id}:
    x-swagger-router-controller: user_controller
    get:
      operationId: checkEmail
      tags:
        - user
      description: update a User
      parameters:
        - name: id
          description: User id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/find/{id}:
    x-swagger-router-controller: user_controller
    get:
      operationId: findUsers
      tags: 
        - user
      description: replace user info
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/User"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  ####################### USERS ########################
  
  ####################### LOGIN ########################
  /login:
    x-swagger-router-controller: authentication_controller
    post:
      operationId: login
      tags: 
        - authentication
      description: Log in user in api
      parameters:
        - name: title
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserLogin"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /category/{id}:
    x-swagger-router-controller: category_controller
    get:
      operationId: getCategory
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: Get category by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    put:
      operationId: updateCategory
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      description: add a category
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: category
          description: Catergory properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: 
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: delete category by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
            
  ################### CATEGORY ####################
  
  ################### GAMES #######################
...
  /game/{id}:
    x-swagger-router-controller: game_controller
    get:
      operationId: getGame
      tags:
        - game
      description: get game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Game"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    post:
      operationId: updateGame
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      summary: Add a new Game
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - in: formData
          name: card_image 
          type: file
          description: The file to upload.
        - in: formData
          name: logo 
          type: file
          description: The file to upload.
        - in: formData
          name: route
          type: string
        - in: formData
          name: name
          type: string 
        - in: formData
          name: description
          type: string
        - in: formData
          name: min_players
          type: integer
        - in: formData
          name: max_players
          type: integer
        - in: formData
          name: color
          type: string
        - in: formData 
          name: category_id
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteGame
      tags: 
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: delete Game by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ApiResponse:
    type: object
    properties:
      success:
        type: boolean
        description: returns true if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  User:
    type: object
    properties:
      id:
        type: string
        description: unique id from user
      username:
        type: string
        description: username from User must be unique
      password: 
        type: string 
        description: password of user can't be less than 6 characters
      email:
        type: string
        description: email of user must be unique
      avatar:
        type: string
        description: image of user
      isAdmin:
        type: boolean
        description: true is admin false is not
      country:
        type: string
      token:
        type: string
        description: user session
      since:
        type: string
        description: date of sign up
  UserProfile:
    type: object
    properties:
      id:
        type: string
        description: unique id from user
      username:
        type: string
        description: username from User must be unique
      email:
        type: string
        description: email of user must be unique
      avatar:
        type: string
        description: image of user
      country:
        type: string
      isFriend:
        type: string
      since:
        type: string
        description: date of sign up
  UserLogin:
    type: object
    properties:
      username:
        type: string
        description: username from User must be unique
      password: 
        type: string 
        description: password of user can't be less than 6 characters
  LoginResponse:
    type: object
    properties:
      success:
        type: boolean
      messages:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            message:
              type: string
      token: 
        type: string
      username:
        type: string
      avatar:
        type: string
      country:
        type: string
      id:
        type: string
  Category:
    type: object
    properties:
      id:
        type: string
      name: 
        type: string 
      description: 
        type: string 
      color:
        type: string
  Game:
    type: object
    properties:
      id:
        type: string
      route: 
        type: string
      name: 
        type: string 
      description: 
        type: string 
      players:
        type: string
      min_players:
        type: integer
      max_players:
        type: integer
      color:
        type: string
      created_at:
        type: string
      card_image:
        type: string
      logo:
        type: string
      category:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          color:
            type: string
...